Move browse filter to it's own component.
